Remove deprecated code from PostCoordinator
